Function Expression body only evaluated if resolved. Changed num to Integer from Double

+++ justin/NumericAtom.java
@@ -1,11 +1,11 @@
 // numeric atom handles parsed numbers.
 // treated as double-precision as just a catch-all.
 public class NumericAtom extends Atom {
-	private double val = 0;
+	private int val = 0;
 	
 	// *************************************************
 	// constructor is used when parsing.
-	public NumericAtom(double val_) {
+	public NumericAtom(int val_) {
 	  val = val_;	
 	
 	}
